requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1

celery
django-celery-beat
redis

celery~=5.4.0